Add exec workaround for sip skipped processes
